moved readFile and readFileAsBytes to test code as they are not used in production
