+++ src/client/java/teammates/client/scripts/GenerateLargeScaledData.java
@@ -6,12 +6,12 @@ import java.io.IOException;
 import teammates.client.remoteapi.RemoteApiClient;
 import teammates.common.datatransfer.DataBundle;
 import teammates.common.datatransfer.FeedbackResponseAttributes;
-import teammates.common.util.FileHelper;
 import teammates.common.util.Utils;
 import teammates.logic.api.Logic;
 import teammates.logic.core.FeedbackQuestionsLogic;
 import teammates.storage.datastore.Datastore;
 import teammates.test.driver.TestProperties;
+import teammates.test.util.FileHelper;
 
 public class GenerateLargeScaledData extends RemoteApiClient {
