+++ src/main/java/teammates/common/util/FileHelper.java
@@ -1,10 +1,5 @@
 package teammates.common.util;
 
-import java.io.BufferedReader;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.util.Scanner;
 
@@ -16,25 +11,6 @@ public final class FileHelper {
         // utility class
     }
     
-    /**
-     * Reads a file content and return a String
-     */
-    public static String readFile(String filename) throws FileNotFoundException {
-        Scanner scanner = new Scanner(new BufferedReader(new FileReader(filename)));
-        String ans = scanner.useDelimiter("\\Z").next();
-        scanner.close();
-        return ans;
-    }
-    
-    public static byte[] readFileAsBytes(String fileName) throws FileNotFoundException, IOException {
-        FileInputStream stream = new FileInputStream(fileName);
-        byte[] buffer = new byte[1024 * 300];
-        stream.read(buffer);
-        stream.close();
-        
-        return buffer;
-    }
-
     /**
      * Reads the contents of a file in the {@code resources} folder
      * as an {@link InputStream}.
