+++ src/test/java/teammates/test/cases/testdriver/HtmlHelperTest.java
@@ -6,10 +6,10 @@ import java.io.IOException;
 
 import org.testng.annotations.Test;
 
-import teammates.common.util.FileHelper;
 import teammates.test.cases.BaseTestCase;
 import teammates.test.driver.HtmlHelper;
 import teammates.test.driver.TestProperties;
+import teammates.test.util.FileHelper;
 
 public class HtmlHelperTest extends BaseTestCase {
