+++ src/test/java/teammates/test/cases/ui/browsertests/InstructorSearchPageUiTest.java
@@ -10,12 +10,12 @@ import teammates.common.datatransfer.DataBundle;
 import teammates.common.datatransfer.StudentAttributes;
 import teammates.common.util.AppUrl;
 import teammates.common.util.Const;
-import teammates.common.util.FileHelper;
 import teammates.common.util.Utils;
 import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.Browser;
 import teammates.test.pageobjects.BrowserPool;
 import teammates.test.pageobjects.InstructorSearchPage;
+import teammates.test.util.FileHelper;
 
 public class InstructorSearchPageUiTest extends BaseUiTestCase {
     private static Browser browser;
