+++ src/test/java/teammates/test/cases/ui/browsertests/MashupPageUiTest.java
@@ -8,12 +8,12 @@ import org.testng.annotations.Test;
 
 import teammates.common.datatransfer.DataBundle;
 import teammates.common.util.Const;
-import teammates.common.util.FileHelper;
 import teammates.common.util.Utils;
 import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.AppPage;
 import teammates.test.pageobjects.Browser;
 import teammates.test.pageobjects.BrowserPool;
+import teammates.test.util.FileHelper;
 
 /**
  * Loads the Mashup page for the tester to do a visual inspection.
