+++ src/test/java/teammates/test/util/FileHelper.java
@@ -1,6 +1,12 @@
 package teammates.test.util;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.IOException;
+import java.util.Scanner;
 
 /**
  * File-related helper methods used for testing. There is another FileHelper on
@@ -12,6 +18,21 @@ public final class FileHelper {
         // utility class
     }
 
+    public static String readFile(String filename) throws FileNotFoundException {
+        Scanner scanner = new Scanner(new BufferedReader(new FileReader(filename)));
+        String ans = scanner.useDelimiter("\\Z").next();
+        scanner.close();
+        return ans;
+    }
+    
+    public static byte[] readFileAsBytes(String fileName) throws IOException {
+        FileInputStream stream = new FileInputStream(fileName);
+        byte[] buffer = new byte[1024 * 300];
+        stream.read(buffer);
+        stream.close();
+        return buffer;
+    }
+    
     public static void deleteFile(String fileName) {
         File file = new File(fileName);
         file.delete();
